BOAC-5251, proper page title; plane-go-round spinner
