+++ src/plugins/vuetify.ts
@@ -17,7 +17,7 @@ import {VCombobox} from 'vuetify/components/VCombobox'
 import {VDataTable, VDataTableVirtual} from 'vuetify/components/VDataTable'
 import {VDialog} from 'vuetify/components/VDialog'
 import {VDivider} from 'vuetify/components/VDivider'
-import {VExpandTransition} from 'vuetify/components/transitions'
+import {VExpandTransition, VFadeTransition} from 'vuetify/components/transitions'
 import {VExpansionPanel, VExpansionPanels, VExpansionPanelText, VExpansionPanelTitle} from 'vuetify/components/VExpansionPanel'
 import {VFileInput} from 'vuetify/components/VFileInput'
 import {VFooter} from 'vuetify/components/VFooter'
@@ -70,6 +70,7 @@ export default createVuetify({
     VExpansionPanels,
     VExpansionPanelText,
     VExpansionPanelTitle,
+    VFadeTransition,
     VFileInput,
     VFooter,
     VIcon,
